refactor: 调整目录结构

+++ backend/src/main/kotlin/com/chengtao/bikesharing/model/Model.kt
@@ -6,7 +6,7 @@ import java.util.Date
  * 共享单车模型
  */
 data class Bike(
-  var id: Short,
+  var id: Int,
   var name: String,
   var foundedAt: Date? = null,
   var createdAt: Date,
@@ -18,7 +18,7 @@ data class Bike(
  */
 data class Development(
   var id: Int,
-  var bikeId: Short,
+  var bikeId: Int,
   var city: String,
   var deliveryAt: Date,
   var createdAt: Date,
